Add extra Testclass for pattern matching with variable mappings
