+++ gradoop-flink/src/test/java/org/gradoop/flink/model/impl/operators/matching/single/preserving/SubgraphIsomorphismTest.java
@@ -1,6 +1,6 @@
 package org.gradoop.flink.model.impl.operators.matching.single.preserving;
 
-import org.gradoop.flink.model.impl.operators.matching.single.PatternMatchingTest;
+import org.gradoop.flink.model.impl.operators.matching.single.PatternMatchingWithBindingTest;
 import org.junit.runners.Parameterized;
 
 import java.util.Arrays;
@@ -11,7 +11,7 @@ import static org.gradoop.flink.model.impl.operators.matching.TestData.*;
  * Test data for pattern matching tests. The graphs are visualized in
  * dev-support/pattern_matching_testcases.pdf
  */
-public abstract class SubgraphIsomorphismTest extends PatternMatchingTest {
+public abstract class SubgraphIsomorphismTest extends PatternMatchingWithBindingTest {
 
   public SubgraphIsomorphismTest(String testName, String dataGraph,
     String queryGraph, String expectedGraphVariables,
